--- NMS/SPI/build.gradle.kts
@@ -1,4 +1,0 @@
-
-dependencies {
-    compileOnly("io.papermc.paper:paper-api:1.21.1-R0.1-SNAPSHOT")
-}
